+++ backend/main.py
@@ -8,7 +8,7 @@ from fastapi.staticfiles import StaticFiles
 
 from db import crear_tablas
 from auth import router as auth
-from router import posts, equipo, events
+from router import posts, equipo, events, estudiantes
 
 
 load_dotenv()
@@ -41,6 +41,7 @@ app.mount("/uploads",StaticFiles(directory="uploads"), name="Uploads")
 app.include_router(posts.router, prefix="/post", tags=["post"])
 app.include_router(events.router, prefix="/event", tags=["event"])
 app.include_router(equipo.router, prefix="/equipo", tags=["equipo"])
+app.include_router(estudiantes.router, prefix="/estudiantes", tags=["estudiantes"])
 app.include_router(auth.router, prefix="/auth", tags=["auth"])
 
 #Main
