+++ backend/router/equipo.py
@@ -4,7 +4,7 @@ from uuid import uuid4
 from fastapi import APIRouter, status, Depends, HTTPException, Form, File, UploadFile
 from typing import Optional
 
-from models.equipo import EstudianteBase, EquipoBase, EstudianteEdit, EquipoEdit, Estudiante, Equipo
+from models.equipo import EquipoBase, EquipoEdit, Equipo
 from models.posts import PostBase, Post, PostsListResponse, PostWithEquipoResponse
 
 from db import db_commit, db_select_query, db_select_range, db_select_unique, db_delete_unique, db_update, select
@@ -15,7 +15,7 @@ from utils import IMG_PATH, IMG_PATH_POSTS, IMG_PATH_USERS
 
 router = APIRouter()
 
-@router.get("/healty")
+@router.post("/healty")
 async def healty(current_team:Equipo=Depends(get_current_user)):
     if not current_team:
         return {"message":"error"}
@@ -26,6 +26,16 @@ async def healty(current_team:Equipo=Depends(get_current_user)):
         "message":"ok",
         "data": data
         }
+
+## VERIFICAR COOKIES
+#
+# from fastapi import Request
+# import logging
+#
+# @router.post("/healty")
+# async def healty(request: Request):
+#     logging.warning(request.cookies)
+#    return {"cookies": request.cookies}
 
 # CRUD Equipo
 @router.post("/crear", status_code=status.HTTP_201_CREATED)
@@ -47,7 +57,7 @@ async def editar(
     if current_team.equipo_id != id:
         raise HTTPException(
             status_code=403,
-            detail="No puedes editar el perfil de otro usuario"
+            detail="No puedes editar el perfil de otro equipo"
         )
     
     # Preparar los datos para actualizar
@@ -88,139 +98,6 @@ async def equipo_borrar(id:int, current_admin:str=Depends(get_current_admin)):
     await db_delete_unique(Equipo,id)
     return {"message":"Equipo eliminado con Exito"}
 
-# CRUD Estudiante
-
-@router.post("/estudiante/crear", status_code=status.HTTP_201_CREATED)
-async def estudiante_crear(
-    name: str = Form(...),
-    codigo: str = Form(...),
-    desc: str | None = Form(None),
-    phone: str | None = Form(None),
-    email: str | None = Form(None),
-    instagram: str | None = Form(None),
-    twiter: str | None = Form(None),
-    tiktok: str | None = Form(None),
-    img: UploadFile = File(...),
-    current_team: Equipo = Depends(get_current_user)
-):  
-    if not img.content_type.startswith("image/"):
-        raise HTTPException(status_code=400, detail="Invalid image type")
-
-    contents = await img.read()
-
-    # Save or process the image
-    filename = f"{uuid4()}.jpg"
-    os.makedirs(IMG_PATH, exist_ok=True)
-    os.makedirs(IMG_PATH_USERS, exist_ok=True)
-    with open(f"{IMG_PATH_USERS}/{filename}", "wb") as f:
-        f.write(contents)
-
-    # Build your DB object manually
-    data = {
-        "name": name,
-        "codigo": codigo,
-        "desc": desc,
-        "phone": phone,
-        "email": email,
-        "instagram": instagram,
-        "twiter": twiter,
-        "tiktok": tiktok,
-        "equipo_id": current_team.equipo_id,
-        "img": f"{IMG_PATH_USERS}/{filename}",
-    }
-
-    est = Estudiante.model_validate(data)
-    await db_commit(est)
-
-    return {"message": "Estudiante creado con éxito"}  
-
-@router.patch("/estudiante/editar/{id}", status_code=status.HTTP_202_ACCEPTED)
-async def estudiante_editar(
-    id: int,
-    name: str = Form(None),
-    codigo: str = Form(None),
-    desc: str = Form(None),
-    phone: str = Form(None),
-    email: str = Form(None),
-    instagram: str = Form(None),
-    twiter: str = Form(None),
-    tiktok: str = Form(None),
-    img: UploadFile = File(None),
-    equipo_id: int = Form(None),
-    current_team: Equipo = Depends(get_current_user)
-):
-    equipo_actual = await db_select_unique(Equipo, id)
-    if current_team.equipo_id != equipo_actual.equipo_id:
-        raise HTTPException(
-            status_code=403,
-            detail="No puedes Editar un Estudiante de otro equipo"
-        )
-
-    # Preparar los datos para actualizar
-    data_dump = {}
-    
-    if name is not None:
-        data_dump["name"] = name
-    if codigo is not None:
-        data_dump["codigo"] = codigo
-    if desc is not None:
-        data_dump["desc"] = desc
-    if phone is not None:
-        data_dump["phone"] = phone
-    if email is not None:
-        data_dump["email"] = email
-    if instagram is not None:
-        data_dump["instagram"] = instagram
-    if twiter is not None:
-        data_dump["twiter"] = twiter
-    if tiktok is not None:
-        data_dump["tiktok"] = tiktok
-    if equipo_id is not None:
-        data_dump["equipo_id"] = equipo_id
-    
-    # Validar que el equipo tenga permiso para editar
-    if equipo_id is not None and current_team.equipo_id != equipo_id:
-        raise HTTPException(
-            status_code=403,
-            detail="No puedes Editar un Estudiante para otro equipo"
-        )
-    
-    # Procesar la imagen si se proporcionó
-    if img is not None:
-        if not img.content_type.startswith("image/"):
-            raise HTTPException(status_code=400, detail="Tipo de imagen inválido")
-        
-        contents = await img.read()
-        
-        # Guardar la imagen
-        filename = f"{uuid4()}.jpg"
-        os.makedirs(IMG_PATH, exist_ok=True)
-        os.makedirs(IMG_PATH_USERS, exist_ok=True)
-        
-        with open(f"{IMG_PATH_USERS}/{filename}", "wb") as f:
-            f.write(contents)
-        
-        data_dump["img"] = f"{IMG_PATH_USERS}/{filename}"
-    
-    # Solo actualizar si hay datos
-    if data_dump:
-        await db_update(Estudiante, id, data_dump)
-    
-    return {"message": "Estudiante editado con éxito"}
-
-@router.delete("/estudiante/delete/{id}", status_code=status.HTTP_403_FORBIDDEN)
-async def estudiante_borrar(id:int, current_team:Equipo=Depends(get_current_user)):
-    est:Estudiante = await db_select_unique(Estudiante,id)
-
-    if current_team.equipo_id != est.equipo_id:
-        raise HTTPException(
-            status_code=403,
-            detail="No puedes Borrar un Estudiante para otro equipo"
-        )
-
-    await db_delete_unique(Estudiante,id)
-    return {"message":"Estudiante eliminado con Exito"}
-
 # GET DATA
 
 @router.get("/get/{id}")
@@ -233,12 +110,6 @@ async def equipo_get_filter(evento_id:int):
     query = select(Equipo).where(Equipo.evento_id == evento_id)
     data = await db_select_query(query)
     return {"data":data,"message":f"Equipos del torneo {evento_id}"}
-
-@router.get("/estudiante/filter/{equipo_id}")
-async def estudiante_get_filter(equipo_id:int):
-    query = select(Estudiante).where(Estudiante.equipo_id == equipo_id)
-    data = await db_select_query(query)
-    return {"data":data,"message":f"Estudiantes del equipo {equipo_id}"}
 
 # POSTS
