+++ frontend/equipos/css/equipos.css
@@ -38,6 +38,7 @@
 }
 
 #modalClose {
+    z-index: 10;
     position: absolute;
     top : 20px;
     right: 20px;
