+++ frontend/equipos/index.html
@@ -59,6 +59,6 @@
 
     <!-- Scripts -->
     <script src="../js/components.js" type="module"></script>
-    <script src="./js/equipos.js"></script>
+    <script src="./js/equipos.js" type="module"></script>
 </body>
 </html>
