+++ frontend/eventos/index.html
@@ -36,6 +36,6 @@
 
     <!-- Scrpts -->
     <script src="../js/components.js" type="module"></script>
-    <script src="./eventos.js"></script>
+    <script src="./eventos.js" type="module"></script>
 </body>
 </html>
